fix: MenuItem accessibility [MDS-599]
