+++ next-docs/public/examples/menuItem/Checkbox.tsx
@@ -6,26 +6,22 @@ const Example = () => {
   const [activeOpt2, setActiveOpt2] = useState(false);
   return (
     <>
-      <div className="w-56">
-        <MenuItem
-          role="сheckbox"
-          isSelected={activeOpt1}
-          onClick={useCallback(() => setActiveOpt1(!activeOpt1), [activeOpt1])}
-        >
-          <MenuItem.Checkbox />
-          <MenuItem.Title>Your value</MenuItem.Title>
-        </MenuItem>
-      </div>
-      <div className="w-56">
-        <MenuItem
-          role="сheckbox"
-          isSelected={activeOpt2}
-          onClick={useCallback(() => setActiveOpt2(!activeOpt2), [activeOpt2])}
-        >
-          <MenuItem.Title>Your value</MenuItem.Title>
-          <MenuItem.Checkbox className="rounded-none shadow-roshi moon-checked:bg-roshi" />
-        </MenuItem>
-      </div>
+      <MenuItem
+        isSelected={activeOpt1}
+        onClick={useCallback(() => setActiveOpt1(!activeOpt1), [activeOpt1])}
+        className="w-56"
+      >
+        <MenuItem.Checkbox />
+        <MenuItem.Title>Your value</MenuItem.Title>
+      </MenuItem>
+      <MenuItem
+        isSelected={activeOpt2}
+        onClick={useCallback(() => setActiveOpt2(!activeOpt2), [activeOpt2])}
+        className="w-56"
+      >
+        <MenuItem.Title>Your value</MenuItem.Title>
+        <MenuItem.Checkbox className="rounded-none shadow-roshi moon-checked:bg-roshi" />
+      </MenuItem>
     </>
   );
 };
