+++ next-docs/public/examples/menuItem/MultiTitle.tsx
@@ -6,46 +6,42 @@ const Example = () => {
   const [active, setActive] = useState(false);
   return (
     <>
-      <div className="w-56">
-        <MenuItem>
-          <span>
-            <ControlsChevronRight className="text-moon-24" />
-          </span>
-          <MenuItem.MultiTitle
-            title="Your value"
-            text={
-              <span>
-                Lorem Ipsum is simply dummy text of the printing and typesetting
-                industry
-              </span>
-            }
-          />
-          <span className="text-moon-12 text-trunks">Meta</span>
-        </MenuItem>
-      </div>
-      <div className="w-56">
-        <MenuItem
-          role="сheckbox"
-          isSelected={active}
-          onClick={useCallback(() => setActive(!active), [active])}
-        >
-          <span className="px-2">
-            <SportBadminton className="text-moon-24" />
-          </span>
-          <MenuItem.MultiTitle
-            title="Your value"
-            text={
-              <span>
-                Lorem Ipsum is simply dummy text of the printing and typesetting
-                industry
-              </span>
-            }
-          />
-          <span className="px-2">
-            <MenuItem.Checkbox />
-          </span>
-        </MenuItem>
-      </div>
+      <MenuItem className="w-56">
+        <span>
+          <ControlsChevronRight className="text-moon-24" />
+        </span>
+        <MenuItem.MultiTitle
+          title="Your value"
+          text={
+            <span>
+              Lorem Ipsum is simply dummy text of the printing and typesetting
+              industry
+            </span>
+          }
+        />
+        <span className="text-moon-12 text-trunks">Meta</span>
+      </MenuItem>
+      <MenuItem
+        isSelected={active}
+        onClick={useCallback(() => setActive(!active), [active])}
+        className="w-56"
+      >
+        <span className="px-2">
+          <SportBadminton className="text-moon-24" />
+        </span>
+        <MenuItem.MultiTitle
+          title="Your value"
+          text={
+            <span>
+              Lorem Ipsum is simply dummy text of the printing and typesetting
+              industry
+            </span>
+          }
+        />
+        <span className="px-2">
+          <MenuItem.Checkbox />
+        </span>
+      </MenuItem>
     </>
   );
 };
