+++ next-docs/public/examples/menuItem/Radio.tsx
@@ -6,26 +6,22 @@ const Example = () => {
   const [activeOpt2, setActiveOpt2] = useState(false);
   return (
     <>
-      <div className="w-56">
-        <MenuItem
-          role="radio"
-          isSelected={activeOpt1}
-          onClick={useCallback(() => setActiveOpt1(!activeOpt1), [activeOpt1])}
-        >
-          <MenuItem.Radio />
-          <MenuItem.Title>Your value</MenuItem.Title>
-        </MenuItem>
-      </div>
-      <div className="w-56">
-        <MenuItem
-          role="radio"
-          isSelected={activeOpt2}
-          onClick={useCallback(() => setActiveOpt2(!activeOpt2), [activeOpt2])}
-        >
-          <MenuItem.Title>Your value</MenuItem.Title>
-          <MenuItem.Radio className="moon-checked:shadow-hit after:bg-hit" />
-        </MenuItem>
-      </div>
+      <MenuItem
+        isSelected={activeOpt1}
+        onClick={useCallback(() => setActiveOpt1(!activeOpt1), [activeOpt1])}
+        className="w-56"
+      >
+        <MenuItem.Radio />
+        <MenuItem.Title>Your value</MenuItem.Title>
+      </MenuItem>
+      <MenuItem
+        isSelected={activeOpt2}
+        onClick={useCallback(() => setActiveOpt2(!activeOpt2), [activeOpt2])}
+        className="w-56"
+      >
+        <MenuItem.Title>Your value</MenuItem.Title>
+        <MenuItem.Radio className="moon-checked:shadow-hit after:bg-hit" />
+      </MenuItem>
     </>
   );
 };
