+++ next-docs/public/examples/menuItem/WithIcon.tsx
@@ -3,18 +3,14 @@ import { SportBadminton } from '@heathmont/moon-icons-tw';
 
 const Example = () => (
   <>
-    <div className="w-56">
-      <MenuItem>
-        <SportBadminton className="text-moon-24" />
-        <MenuItem.Title>Your value</MenuItem.Title>
-      </MenuItem>
-    </div>
-    <div className="w-56">
-      <MenuItem>
-        <MenuItem.Title>Your value</MenuItem.Title>
-        <SportBadminton className="text-moon-24" />
-      </MenuItem>
-    </div>
+    <MenuItem className="w-56">
+      <SportBadminton className="text-moon-24" />
+      <MenuItem.Title>Your value</MenuItem.Title>
+    </MenuItem>
+    <MenuItem className="w-56">
+      <MenuItem.Title>Your value</MenuItem.Title>
+      <SportBadminton className="text-moon-24" />
+    </MenuItem>
   </>
 );
