+++ next-docs/public/examples/menuItem/WithMeta.tsx
@@ -3,19 +3,15 @@ import { ControlsChevronRight } from '@heathmont/moon-icons-tw';
 
 const Example = () => (
   <>
-    <div className="w-56">
-      <MenuItem>
-        <MenuItem.Title>Your value</MenuItem.Title>
-        <span className="text-moon-12 text-trunks">Meta</span>
-      </MenuItem>
-    </div>
-    <div className="w-56">
-      <MenuItem>
-        <ControlsChevronRight className="text-moon-24" />
-        <MenuItem.Title>Your value</MenuItem.Title>
-        <span className="text-moon-12 text-trunks">Meta</span>
-      </MenuItem>
-    </div>
+    <MenuItem className="w-56">
+      <MenuItem.Title>Your value</MenuItem.Title>
+      <span className="text-moon-12 text-trunks">Meta</span>
+    </MenuItem>
+    <MenuItem className="w-56">
+      <ControlsChevronRight className="text-moon-24" />
+      <MenuItem.Title>Your value</MenuItem.Title>
+      <span className="text-moon-12 text-trunks">Meta</span>
+    </MenuItem>
   </>
 );
