+++ workspaces/core/src/menuItem/private/types/CheckboxRadioProps.ts
@@ -1,0 +1,7 @@
+type CheckboxRadioProps = {
+  isSelected?: boolean;
+  className?: string;
+  ['aria-label']?: string;
+};
+
+export default CheckboxRadioProps;
