+++ workspaces/core/src/menuItem/private/types/MenuItemComponentProps.ts
@@ -1,0 +1,7 @@
+import type MenuItemPolymorphicProps from './MenuItemPolymorphicProps';
+
+type MenuItemComponentProps = <C extends React.ElementType = 'button'>(
+  props: MenuItemPolymorphicProps<C>
+) => React.ReactElement | null;
+
+export default MenuItemComponentProps;
