+++ workspaces/core/src/menuItem/private/types/MenuItemPolymorphicProps.ts
@@ -1,0 +1,6 @@
+import type MenuItemProps from './MenuItemProps';
+
+type MenuItemPolymorphicProps<C extends React.ElementType> =
+  PolymorphicComponentPropWithRef<C, MenuItemProps>;
+
+export default MenuItemPolymorphicProps;
