+++ workspaces/core/src/menuItem/private/types/MenuItemProps.ts
@@ -1,0 +1,8 @@
+type MenuItemProps = {
+  width?: string;
+  isSelected?: boolean;
+  isActive?: boolean;
+  className?: string;
+};
+
+export default MenuItemProps;
