+++ workspaces/core/src/menuItem/private/types/MenuItemState.ts
@@ -1,0 +1,7 @@
+type MenuItemState = {
+  selected?: boolean;
+  active?: boolean;
+  registerChild?: (child: string) => () => void;
+};
+
+export default MenuItemState;
