+++ workspaces/core/src/menuItem/private/types/MultiTitleProps.ts
@@ -1,0 +1,6 @@
+type MultiTitleProps = {
+  title?: JSX.Element | string;
+  text?: JSX.Element | string;
+};
+
+export default MultiTitleProps;
