+++ workspaces/core/src/menuItem/private/utils/MenuItemContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+import type MenuItemState from '../types/MenuItemState';
+
+const MenuItemContext = createContext<MenuItemState>({});
+MenuItemContext.displayName = 'MenuItemContext';
+
+export default MenuItemContext;
