+++ workspaces/core/src/menuItem/private/utils/useMenuItemContext.ts
@@ -1,0 +1,16 @@
+import { useContext } from 'react';
+import MenuItemContext from './MenuItemContext';
+
+const useMenuItemContext = (component: string) => {
+  const context = useContext(MenuItemContext);
+  if (context === null) {
+    const err = new Error(
+      `<${component}> is missing a parent <Menu /> component.`
+    );
+    // if (Error.captureStackTrace) Error.captureStackTrace(err, useMenuItemContext);
+    throw err;
+  }
+  return context;
+};
+
+export default useMenuItemContext;
